+++ test/EaselShape.spec.js
@@ -3,10 +3,14 @@ import EaselCanvas from '../src/components/EaselCanvas.vue';
 import EaselShape from '../src/components/EaselShape.vue';
 import Vue from 'vue';
 import isADisplayObject from './includes/is-a-display-object.js';
+import isAlignable from './includes/is-alignable.js';
+
 
 describe('EaselShape', function () {
 
-    describe('is a display object that', isADisplayObject(EaselShape, 'form="circle"'));
+    describe('is a display object that', isADisplayObject(EaselShape, 'form="circle" dimensions="50"'));
+
+    describe('is alignable and', isAlignable(EaselShape, {width: 100, height: 100}, 'form="circle" dimensions="50"'));
 
     const buildVm = function () {
         const vm = new Vue({
